Fix: cleanup
