+++ spring-webmvc/src/main/java/org/springframework/web/servlet/config/MvcNamespaceUtils.java
@@ -35,7 +35,6 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.servlet.DispatcherServlet;
 import org.springframework.web.servlet.handler.AbstractHandlerMapping;
 import org.springframework.web.servlet.handler.BeanNameUrlHandlerMapping;
-import org.springframework.web.servlet.handler.HandlerMappingIntrospector;
 import org.springframework.web.servlet.i18n.AcceptHeaderLocaleResolver;
 import org.springframework.web.servlet.mvc.HttpRequestHandlerAdapter;
 import org.springframework.web.servlet.mvc.SimpleControllerHandlerAdapter;
@@ -72,14 +71,11 @@ public abstract class MvcNamespaceUtils {
 
 	private static final String CORS_CONFIGURATION_BEAN_NAME = "mvcCorsConfigurations";
 
-	private static final String HANDLER_MAPPING_INTROSPECTOR_BEAN_NAME = "mvcHandlerMappingIntrospector";
-
 
 	public static void registerDefaultComponents(ParserContext context, @Nullable Object source) {
 		registerBeanNameUrlHandlerMapping(context, source);
 		registerHttpRequestHandlerAdapter(context, source);
 		registerSimpleControllerHandlerAdapter(context, source);
-		registerHandlerMappingIntrospector(context, source);
 		registerLocaleResolver(context, source);
 		registerViewNameTranslator(context, source);
 		registerFlashMapManager(context, source);
@@ -273,22 +269,6 @@ public abstract class MvcNamespaceUtils {
 			corsDef.getConstructorArgumentValues().addIndexedArgumentValue(0, corsConfigurations);
 		}
 		return new RuntimeBeanReference(CORS_CONFIGURATION_BEAN_NAME);
-	}
-
-	/**
-	 * Registers an {@link HandlerMappingIntrospector} under a well-known name
-	 * unless already registered.
-	 */
-	@SuppressWarnings("removal")
-	private static void registerHandlerMappingIntrospector(ParserContext context, @Nullable Object source) {
-		if (!context.getRegistry().containsBeanDefinition(HANDLER_MAPPING_INTROSPECTOR_BEAN_NAME)) {
-			RootBeanDefinition beanDef = new RootBeanDefinition(HandlerMappingIntrospector.class);
-			beanDef.setSource(source);
-			beanDef.setRole(BeanDefinition.ROLE_INFRASTRUCTURE);
-			beanDef.setLazyInit(true);
-			context.getRegistry().registerBeanDefinition(HANDLER_MAPPING_INTROSPECTOR_BEAN_NAME, beanDef);
-			context.registerComponent(new BeanComponentDefinition(beanDef, HANDLER_MAPPING_INTROSPECTOR_BEAN_NAME));
-		}
 	}
 
 	/**
