+++ spring-webmvc/src/main/java/org/springframework/web/servlet/handler/RequestMatchResult.java
@@ -34,7 +34,7 @@ import org.springframework.web.util.pattern.PathPatternParser;
  *
  * @author Rossen Stoyanchev
  * @since 4.3.1
- * @deprecated together with {@link HandlerMappingIntrospector} without a replacement.
+ * @deprecated without a replacement.
  */
 @SuppressWarnings("removal")
 @Deprecated(since = "7.0", forRemoval = true)
